Fix attempting to access status values on FALSE
In PHP < 7.4, `var_dump(FALSE['anything'])` gives NULL
without triggering a warning / notice. More recent PHP
versions will not allow that.

`$version` here will be FALSE if the migration is missing
from the list, in which case the status is expected to show
empty values for the start and end time.