Drop support for seeding the database
This is always a development concern, not a production one, so merging the
concept into the database migration tool feels wrong. Removing the database
mapping layer has broken some seed functionality (and that was not caught by
the tests, which are much lighter for seeding than they are for migrating).

Other data-generation/seeding libraries are available and more appropriate
to use as a standalone tool in dev / CI if required.