Remove support for proxying database adapters through a 'wrapper'
The timed adapter was the only remaining implementation of this proxy pattern,
and was only actually printing statistics for `insert()` operations.

This was not that useful, and the overhead (e.g. on identifying dead code /
improving type safety / adding to adapters in future) of the proxy pattern was
quite high relative to the benefit. You can, of course, implement and register
your own adapter class as a proxy if you need it.